+++ src/quickapp/skills/skills_module.py
@@ -5,6 +5,10 @@ from injector import AssistedBuilder, Binder, Module, multiprovider, singleton
 
 from quickapp.common import StagedBaseTool
 from quickapp.common.abstract.base_prompt_provider import PromptPartProvider
+from quickapp.common.abstract.base_transformer import MessagesTransformer
+from quickapp.skills._inject_file_transfer_instruction_transformer import (
+    _InjectFileTransferInstructionTransformer,
+)
 from quickapp.skills._skill_reader_tool import _SkillReaderTool
 from quickapp.skills._tool_configs import SKILL_READER_TOOL_CONFIG, SKILL_READER_TOOL_NAME
 from quickapp.skills.agent_skills_provider import AgentSkillsProvider
@@ -17,6 +21,11 @@ class SkillsModule(Module):
     def configure(self, binder: Binder) -> None:
         binder.bind(AgentSkillsProvider, to=AgentSkillsProvider, scope=singleton)
         binder.bind(_SkillReaderTool, to=_SkillReaderTool, scope=request_scope)
+        binder.bind(
+            _InjectFileTransferInstructionTransformer,
+            to=_InjectFileTransferInstructionTransformer,
+            scope=request_scope,
+        )
 
     @multiprovider
     def _provide_internal_tools(
@@ -37,3 +46,10 @@ class SkillsModule(Module):
         skills_provider: AgentSkillsProvider,
     ) -> list[PromptPartProvider]:
         return [skills_provider]
+
+    @multiprovider
+    def _provide_message_transformers(
+        self,
+        file_transfer_transformer: _InjectFileTransferInstructionTransformer,
+    ) -> list[MessagesTransformer]:
+        return [file_transfer_transformer]
